Don't use private `date_` property (see #1026)
It looks like at some point the Closure Library exposed a public `date` property for this same purpose.  The version of the Closure Library that comes with the closure-util package uses this property.  At some point later, I'll demonstrate how we can control which version of the library we depend on.