Add first cut of MapGuide support in ol3. This adds a new ol.source.MapGuide class that requests single-tiled layers using either the GETMAPIMAGE operation (useOverlay = false) or GETDYNAMICMAPOVERLAYIMAGE operation (useOverlay = true).
When useOverlay = false, you can pass either a Map Definition ID or a session/mapname pair. When useOverlay = true, you must pass a session/mapname pair. Both modes also require a meters-per-unit value to be specified in order for scale calculations to be correct. This value can be calculated with the MapGuide Web API, or through the CREATERUNTIMEMAP operation introduced in MapGuide Open Source 2.6