Add support for indicating API stability with inline documentation
In preparation for documenting the stability of API features, this commit adds support for indicating API stability via a new jsdoc tag @stability.  There is a small plugin for jsdoc that adds a new tag, @stabiility, with some ability to enforce specific level names.  The templates and css have been updated to include the stability level in the generated documentation.