Add an alternative algorithm for zoomify tier size calculation
Zoomify uses a different algorithm than ol3 to calculate the number of
tiles per tier.
These algorithms are compatible in most cases, but they will lead to
different results with some particular image sizes.

This update provides an option for the zoomify source to use one of
these two algorithms.