Raise errors inside methods
The JsDoc states that the methods in this commit must return some value.
Other packages (e.g. plovr) may however show a warning, as the method
itself does not always return a value or throw an error. The package may
be incapable of detecting it's the function call that throws the error.