Fix SVG hit detection on IE
IE does not know the size of SVG icons. When creating a 'hit detection icon',
OpenLayers will fail, as the canvas won't accept a 0x0 sized image.
This commit ensures the 'hit detection icon' is creating using the same size
as the original image. Note that one must provide a size in `ol.style.Icon`
for this to have any effect on IE.