Make source map usable by default
Currently there's a source map generated for `ol.js` but it can't be
used easily. With tweaking the build system, the source files are now
expected to be in a directory called `src` which is on the same level
as the `build` directory where the `ol.js` resides.

This makes local development easier. You can have a local server pointing
to an OpenLayers checkout and you just need to include the `build/ol.js`
file in your project and it will automatically pick up the source map.

The `sourceMappingURL` is only added to the `ol.js` if an `ol.js.map` is
located in the same directory. This means that it will only be added
in a local build, but not when the distribution is created.