Upgrade to geodesicflatgeom.js
Changes in preparation for an upgraded graticule that will be able to
show grids of any projection, not just a lat/lon EPSG:4326 grid.

Make the basic path interpolate and transform function public. It will
be used for generating sets of points a fixed horizontal distance apart
on EPSG:3857. The function is good for transforming 'smooth' paths
between projections.

Add public functions for drawing horizontal and vertical grid lines in
one projection in another projection where they may be curves.