Use --ssl-protocol=any in PhantomJS with a page that uses HTTPS
The `check-examples` target fails on Travis without this option.  We get "SSL handshake failed" when loading resources via HTTPS with whatever the default SSL protocol is.

In addition, as a precaution, the --ignore-ssl-errors option is set to true in case anybody adds a resource with an invalid certificate.  This option does not get around SSL handshake errors.