Simplify tileLoadFunction and crossOrigin handling
It will be easier to work with tile classes other than ol.ImageTile when
crossOrigin and tileLoadFunction are optional constructor parameters. In
the case of the tileLoadFunction, the default should be provided by the
tile class, not the tile source. And in the case of crossOrigin, providing
null as value does not make sense, except for the OSM source, where users
must be able to unset the 'anonymous' default.